# content/imvertor/20251118-import-logical-name/openapi/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "JZV",
    "description" : "Dit is de tekst van de tag *OA Description* bij de package «Informatiemodel» JZV. Dit is een [link](https://www.justid.nl/).",
    "version" : "1.0.0"
  },
...
  "paths" : {
...
    "/v1/casus" : {
      "get" : {
        "tags" : [ "Casus" ],
        "summary" : "Retourneert de lijst van alle Casus objecten",
        "description" : "Retourneert een gepagineerde lijst van alle Casus objecten",
        "operationId" : "haalAlleCasusOp",
        "parameters" : [ {
          "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-parameters.yaml#/components/parameters/page"
        }, {
          "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-parameters.yaml#/components/parameters/pageSize"
        }, {
          "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-parameters.yaml#/components/parameters/sortField"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "headers" : {
              "RateLimit-Remaining" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Remaining",
                "style" : "simple"
              },
              "Warning" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/Warning",
                "style" : "simple"
              },
              "RateLimit-Reset" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Reset",
                "style" : "simple"
              },
              "X-Pagination-Limit" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/X-Pagination-Limit",
                "style" : "simple"
              },
              "X-Pagination-Page" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/X-Pagination-Page",
                "style" : "simple"
              },
              "RateLimit-Limit" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Limit",
                "style" : "simple"
              },
              "API-Version" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/API-Version",
                "style" : "simple"
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GepagineerdCasusLijst"
                }
              }
            }
          },
          "400" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/400"
          },
          "401" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/401"
          },
          "403" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/403"
          },
          "405" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/405"
          },
          "415" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/415"
          },
          "429" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/429"
          },
          "500" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/500"
          },
          "503" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/503"
          }
        }
      },
      "post" : {
        "tags" : [ "Casus" ],
        "summary" : "Maakt een nieuw Casus object",
        "description" : "Maakt een nieuw Casus object aan op basis van de aangeleverde gegevens",
        "operationId" : "maakCasusAan",
        "requestBody" : {
          "description" : "De gegevens van het Casus object",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Casus"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Casus succesvol aangemaakt",
            "headers" : {
              "RateLimit-Remaining" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Remaining",
                "style" : "simple"
              },
              "Warning" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/Warning",
                "style" : "simple"
              },
              "RateLimit-Reset" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Reset",
                "style" : "simple"
              },
              "X-Pagination-Limit" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/X-Pagination-Limit",
                "style" : "simple"
              },
              "X-Pagination-Page" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/X-Pagination-Page",
                "style" : "simple"
              },
              "RateLimit-Limit" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Limit",
                "style" : "simple"
              },
              "API-Version" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/API-Version",
                "style" : "simple"
              },
              "Location" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/Location",
                "style" : "simple"
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Casus"
                }
              }
            }
          },
          "400" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/400"
          },
          "401" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/401"
          },
          "403" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/403"
          },
          "405" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/405"
          },
          "409" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/409"
          },
          "415" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/415"
          },
          "422" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/422"
          },
          "429" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/429"
          },
          "500" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/500"
          },
          "503" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/503"
          }
        }
      }
    },
...
    "/v1/casusoverleg" : {
      "get" : {
        "tags" : [ "Casusoverleg" ],
        "summary" : "Retourneert de lijst van alle Casusoverleg objecten",
        "description" : "Retourneert een gepagineerde lijst van alle Casusoverleg objecten",
        "operationId" : "haalAlleCasusoverlegOp",
        "parameters" : [ {
          "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-parameters.yaml#/components/parameters/page"
        }, {
          "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-parameters.yaml#/components/parameters/pageSize"
        }, {
          "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-parameters.yaml#/components/parameters/sortField"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "headers" : {
              "RateLimit-Remaining" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Remaining",
                "style" : "simple"
              },
              "Warning" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/Warning",
                "style" : "simple"
              },
              "RateLimit-Reset" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Reset",
                "style" : "simple"
              },
              "X-Pagination-Limit" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/X-Pagination-Limit",
                "style" : "simple"
              },
              "X-Pagination-Page" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/X-Pagination-Page",
                "style" : "simple"
              },
              "RateLimit-Limit" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Limit",
                "style" : "simple"
              },
              "API-Version" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/API-Version",
                "style" : "simple"
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GepagineerdCasusoverlegLijst"
                }
              }
            }
          },
          "400" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/400"
          },
          "401" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/401"
          },
          "403" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/403"
          },
          "405" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/405"
          },
          "415" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/415"
          },
          "429" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/429"
          },
          "500" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/500"
          },
          "503" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/503"
          }
        }
      },
      "post" : {
        "tags" : [ "Casusoverleg" ],
        "summary" : "Maakt een nieuw Casusoverleg object",
        "description" : "Maakt een nieuw Casusoverleg object aan op basis van de aangeleverde gegevens",
        "operationId" : "maakCasusoverlegAan",
        "requestBody" : {
          "description" : "De gegevens van het Casusoverleg object",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Casusoverleg"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Casusoverleg succesvol aangemaakt",
            "headers" : {
              "RateLimit-Remaining" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Remaining",
                "style" : "simple"
              },
              "Warning" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/Warning",
                "style" : "simple"
              },
              "RateLimit-Reset" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Reset",
                "style" : "simple"
              },
              "X-Pagination-Limit" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/X-Pagination-Limit",
                "style" : "simple"
              },
              "X-Pagination-Page" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/X-Pagination-Page",
                "style" : "simple"
              },
              "RateLimit-Limit" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Limit",
                "style" : "simple"
              },
              "API-Version" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/API-Version",
                "style" : "simple"
              },
              "Location" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/Location",
                "style" : "simple"
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Casusoverleg"
                }
              }
            }
          },
          "400" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/400"
          },
          "401" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/401"
          },
          "403" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/403"
          },
          "405" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/405"
          },
          "409" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/409"
          },
          "415" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/415"
          },
          "422" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/422"
          },
          "429" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/429"
          },
          "500" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/500"
          },
          "503" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/503"
          }
        }
      }
    },
    "/v1/casusoverleg-datum" : {
      "get" : {
        "tags" : [ "Casusoverleg" ],
        "summary" : "Zoek overleg op datum",
        "description" : "Retourneert de Overleggen die hebben plaatsgevonden op de gespecificeerde datum.",
        "operationId" : "vindCasusoverlegOpDatum",
        "parameters" : [ {
          "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-parameters.yaml#/components/parameters/page"
        }, {
          "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-parameters.yaml#/components/parameters/pageSize"
        }, {
          "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-parameters.yaml#/components/parameters/sortField"
        }, {
          "name" : "zoekdatum",
          "in" : "query",
          "description" : "De datum waarop het overleg heeft plaatsgevonden.",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "date"
          },
          "example" : "2025-10-01"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "headers" : {
              "RateLimit-Remaining" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Remaining",
                "style" : "simple"
              },
              "Warning" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/Warning",
                "style" : "simple"
              },
              "RateLimit-Reset" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Reset",
                "style" : "simple"
              },
              "X-Pagination-Limit" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/X-Pagination-Limit",
                "style" : "simple"
              },
              "X-Pagination-Page" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/X-Pagination-Page",
                "style" : "simple"
              },
              "RateLimit-Limit" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/RateLimit-Limit",
                "style" : "simple"
              },
              "API-Version" : {
                "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-headers.yaml#/components/headers/API-Version",
                "style" : "simple"
              }
            },
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GepagineerdCasusoverlegLijst"
                }
              }
            }
          },
          "400" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/400"
          },
          "401" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/401"
          },
          "403" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/403"
          },
          "405" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/405"
          },
          "415" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/415"
          },
          "429" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/429"
          },
          "500" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/500"
          },
          "503" : {
            "$ref" : "https://armatiek.github.io/imvertor-openapi-generator/openapi/common-responses.yaml#/components/responses/503"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Afspraak" : {
        "required" : [ "id", "omschrijving", "url", "verantwoordelijkeOrganisatie" ],
        "type" : "object",
        "properties" : {
          "url" : {
            "minLength" : 1,
            "type" : "string",
            "description" : "URL-referentie naar de resource waarnaar verwezen wordt",
            "format" : "uri",
            "readOnly" : true
          },
          "id" : {
            "type" : "string",
            "description" : "De unieke identificatie van de afspraak."
          },
          "omschrijving" : {
            "type" : "string",
            "description" : "De beschrijving van de gemaakte afspraak."
          },
          "verantwoordelijkeOrganisatie" : {
            "type" : "string"
          }
        },
        "description" : "Een toezegging van een ketenpartner bepaalde handelingen of interventies zo snel mogelijk uit te gaan voeren.",
        "nullable" : true
      },
      "Besluit" : {
        "required" : [ "id", "omschrijving", "url" ],
        "type" : "object",
        "properties" : {
          "url" : {
            "minLength" : 1,
            "type" : "string",
            "description" : "URL-referentie naar de resource waarnaar verwezen wordt",
            "format" : "uri",
            "readOnly" : true
          },
          "id" : {
            "type" : "string"
          },
          "omschrijving" : {
            "type" : "string"
          }
        },
        "nullable" : true
      },
      "Betrokkene" : {
        "required" : [ "burgerservicenummer", "naam", "url" ],
        "type" : "object",
        "properties" : {
          "url" : {
            "minLength" : 1,
            "type" : "string",
            "description" : "URL-referentie naar de resource waarnaar verwezen wordt",
            "format" : "uri",
            "readOnly" : true
          },
          "burgerservicenummer" : {
            "type" : "string",
            "description" : "Het burgerservicenummer van de betrokkene."
          },
          "naam" : {
            "$ref" : "#/components/schemas/Naam"
          }
        },
        "description" : "Een persoon die onderwerp is van de casus."
      },
      "Casus" : {
        "required" : [ "heeft hoofdbetrokkene", "id", "startDatum", "url" ],
        "type" : "object",
        "properties" : {
          "url" : {
            "minLength" : 1,
            "type" : "string",
            "description" : "URL-referentie naar de resource waarnaar verwezen wordt",
            "format" : "uri",
            "readOnly" : true
          },
          "id" : {
            "type" : "string"
          },
          "startDatum" : {
            "type" : "string",
            "format" : "date"
          },
          "heeft nevenbetrokkene" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Betrokkene"
            }
          },
          "heeftScenario" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Scenario"
            }
          },
          "heeftAfspraak" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Afspraak"
            }
          },
          "heeftBesluit" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Besluit"
            }
          },
          "heeft hoofdbetrokkene" : {
            "$ref" : "#/components/schemas/Betrokkene"
          }
        }
      },
      "Casusoverleg" : {
        "required" : [ "datum", "heeft agendapunt", "id", "notulen", "url" ],
        "type" : "object",
        "properties" : {
          "url" : {
            "minLength" : 1,
            "type" : "string",
            "description" : "URL-referentie naar de resource waarnaar verwezen wordt",
            "format" : "uri",
            "readOnly" : true
          },
          "id" : {
            "type" : "string",
            "description" : "De unieke identificatie van het casusoverleg."
          },
          "datum" : {
            "type" : "string",
            "description" : "De datum waarop het casusoverleg heeft plaatsgevonden.",
            "format" : "date"
          },
          "notulen" : {
            "type" : "string"
          },
          "heeft agendapunt" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Casus"
            }
          }
        },
        "description" : "Een periodiek beraad waarbij meerdere bij de casus betrokken ketenpartners in een samenwerking samenwerken.\n\nZie deze [link](https://www.justid.nl/) voor meer informatie."
      },
      "GepagineerdBase" : {
        "required" : [ "pagina", "paginaGrootte" ],
        "type" : "object",
        "properties" : {
          "pagina" : {
            "minimum" : 1,
            "type" : "integer",
            "description" : "Huidig paginanummer",
            "format" : "int32"
          },
          "paginaGrootte" : {
            "maximum" : 100,
            "type" : "integer",
            "description" : "Aantal objecten per pagina",
            "format" : "int32"
          },
          "totaalAantal" : {
            "minimum" : 0,
            "type" : "integer",
            "description" : "Totaal aantal beschikbare objecten",
            "format" : "int32",
            "nullable" : true
          },
          "sorteer" : {
            "type" : "string",
            "description" : "Komma gescheiden lijst van velden waarop gesorteerd moet worden. Door een minteken (-) voor de veldnaam te zetten wordt het veld in aflopende sorteervolgorde gesorteerd.",
            "example" : "name,-geboorteDatum"
          },
          "volgende" : {
            "type" : "string",
            "description" : "Link naar de volgende pagina",
            "format" : "uri",
            "nullable" : true,
            "example" : "http://mijn-api.org/api/v1/klanten/?pagina=4&paginaGrootte=10&sorteer=name,-geboorteDatum"
          },
          "vorige" : {
            "type" : "string",
            "description" : "Link naar de vorige pagina",
            "format" : "uri",
            "nullable" : true,
            "example" : "http://mijn-api.org/api/v1/klanten/?pagina=2&paginaGrootte=10&sorteer=name,-geboorteDatum"
          }
        },
        "description" : "Basis schema voor gepagineerde resultaten"
      },
      "GepagineerdCasusoverlegLijst" : {
        "required" : [ "resultaten" ],
        "type" : "object",
        "allOf" : [ {
          "$ref" : "#/components/schemas/GepagineerdBase"
        }, {
          "type" : "object",
          "properties" : {
            "resultaten" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/Casusoverleg"
              }
            }
          }
        } ]
      },
      "Naam" : {
        "required" : [ "Geslachtsnaam", "Tussenvoegsel", "Voornaam", "id", "url" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "minLength" : 1,
            "type" : "string",
            "description" : "Unieke identificatie van de resource waarnaar verwezen wordt",
            "readOnly" : true
          },
          "url" : {
            "minLength" : 1,
            "type" : "string",
            "description" : "URL-referentie naar de resource waarnaar verwezen wordt",
            "format" : "uri",
            "readOnly" : true
          },
          "Voornaam" : {
            "type" : "string"
          },
          "Tussenvoegsel" : {
            "type" : "string"
          },
          "Geslachtsnaam" : {
            "type" : "string"
          }
        },
        "description" : "Gegevensgroeptype voor naamgegevens."
      },
      "Scenario" : {
        "required" : [ "id", "omschrijving", "url", "verantwoordelijkeOrganisatie" ],
        "type" : "object",
        "properties" : {
          "url" : {
            "minLength" : 1,
            "type" : "string",
            "description" : "URL-referentie naar de resource waarnaar verwezen wordt",
            "format" : "uri",
            "readOnly" : true
          },
          "id" : {
            "type" : "string"
          },
          "omschrijving" : {
            "type" : "string"
          },
          "verantwoordelijkeOrganisatie" : {
            "type" : "string"
          }
        }
      },
...
      "GepagineerdCasusLijst" : {
        "required" : [ "resultaten" ],
        "type" : "object",
        "allOf" : [ {
          "$ref" : "#/components/schemas/GepagineerdBase"
        }, {
          "type" : "object",
          "properties" : {
            "resultaten" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/components/schemas/Casus"
              }
            }
          }
        } ]
      },
...
    }
  }
}